Sharing Arduino2 messages between functions
I noticed that if GRBL gets end stop signal when not moving, it probably has some errors in further operatiion.  So I'm changing the way the encoder homing works.
The thread receiving messages from Arduino2 will write the last message to a global variable in settings.py, so refer function can access it.